requirements.txt
flask==3.1.*
flask-login
pytesseract
Pillow
pymupdf
anthropic
psycopg2-binary
python-dotenv
gunicorn
requests
